+++ src/app/auth/forgot-password/forgot-password.action.ts
@@ -1,0 +1,21 @@
+'use server';
+
+import type { ForgotPasswordForm } from '@/schemas/auth/forgot-password.schema';
+import { createClient } from '@/lib/supabase/server';
+import { createLogEvent } from '@/lib/supabase/log-event';
+
+export async function forgotPassword(formData: ForgotPasswordForm) {
+  const supabase = await createClient();
+
+  const { error } = await supabase.auth.resetPasswordForEmail(formData.email);
+
+  if (error) {
+    createLogEvent('error', 'LOST_PASSWORD_FAILED', error.message + '. Email: ' + formData.email);
+
+    return { success: false };
+  }
+
+  createLogEvent('info', 'LOST_PASSWORD_SUCCESSFUL', 'Lost password. Email: ' + formData.email);
+
+  return { success: true };
+}