+++ CHANGELOG.md
@@ -13,7 +13,8 @@ All notable changes to this project will be documented in this file.
 - login form hook using react hook form
 - forgot password form hook using react hook form
 - reset password form hook using react hook form
-- login form sign in action using supabase auth
+- login form action using supabase auth
+- forgot password form action using supabase auth
 
 ##### Dashboard
 
@@ -44,6 +45,7 @@ All notable changes to this project will be documented in this file.
 - forgot password form link to login to include auth
 - update password form file name to reset password form
 - login form hook to use tanstack query
+- forgot password form hook to use tanstack query
 
 #### Components